pyproject.toml
[project]
name = "softdesk"
version = "0.1.0"
description = ""
authors = [
    {name = "Jules",email = "jules@gmail.com"}
]
readme = "README.md"
packages = [
    { include = "softdesk" }
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "drf-yasg (>=1.21.8,<2.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "cachecontrol (>=0.14.2,<0.15.0)",
    "faker (>=37.0.2,<38.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "rapidfuzz (>=3.12.2,<4.0.0)",
    "anyio (>=4.9.0,<5.0.0)",
    "asgiref (>=3.8.1,<4.0.0)",
    "backports-tarfile (>=1.2.0,<2.0.0)",
    "build (>=1.2.2.post1,<2.0.0)",
    "certifi (>=2025.1.31,<2026.0.0)",
    "cffi (>=1.17.1,<2.0.0)",
    "charset-normalizer (>=3.4.1,<4.0.0)",
    "cleo (>=2.1.0,<3.0.0)",
    "crashtest (>=0.4.1,<0.5.0)",
    "distlib (>=0.3.9,<0.4.0)",
    "dulwich (>=0.22.8,<0.23.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "fastjsonschema (>=2.21.1,<3.0.0)",
    "filelock (>=3.18.0,<4.0.0)",
    "findpython (>=0.6.3,<0.7.0)",
    "h11 (>=0.14.0,<0.15.0)",
    "httpcore (>=1.0.7,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "idna (>=3.10,<4.0)",
    "importlib-metadata (>=8.6.1,<9.0.0)",
    "inflection (>=0.5.1,<0.6.0)",
    "iniconfig (>=2.1.0,<3.0.0)",
    "installer (>=0.7.0,<0.8.0)",
    "jaraco-classes (>=3.4.0,<4.0.0)",
    "jaraco-context (>=6.0.1,<7.0.0)",
    "jaraco-functools (>=4.1.0,<5.0.0)",
    "keyring (>=25.6.0,<26.0.0)",
    "more-itertools (>=10.6.0,<11.0.0)",
    "msgpack (>=1.1.0,<2.0.0)",
    "packaging (>=24.2,<25.0)",
    "pbs-installer (>=2025.3.17,<2026.0.0)",
    "pkginfo (>=1.12.1.2,<2.0.0.0)",
    "platformdirs (>=4.3.7,<5.0.0)",
    "pluggy (>=1.5.0,<2.0.0)",
    "poetry (>=2.1.1,<3.0.0)",
    "poetry-core (>=2.1.1,<3.0.0)",
    "pycparser (>=2.22,<3.0)",
    "pyproject-hooks (>=1.2.0,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-django (>=4.10.0,<5.0.0)",
    "pytest-factoryboy (>=2.7.0,<3.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "pytz (>=2025.1,<2026.0)",
    "requests (>=2.32.3,<3.0.0)",
    "requests-toolbelt (>=1.0.0,<2.0.0)",
    "shellingham (>=1.5.4,<2.0.0)",
    "six (>=1.17.0,<2.0.0)",
    "sniffio (>=1.3.1,<2.0.0)",
    "sqlparse (>=0.5.3,<0.6.0)",
    "tomlkit (>=0.13.2,<0.14.0)",
    "trove-classifiers (>=2025.3.19.19,<2026.0.0.0)",
    "typing-extensions (>=4.12.2,<5.0.0)",
    "uritemplate (>=4.1.1,<5.0.0)",
    "urllib3 (>=2.3.0,<3.0.0)",
    "virtualenv (>=20.29.3,<21.0.0)",
    "xattr (>=1.1.4,<2.0.0)",
    "zipp (>=3.21.0,<4.0.0)",
    "zstandard (>=0.23.0,<0.24.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.9.0"
pytest-factoryboy = "^2.7.0"
factory-boy = "^3.3.3"

